tt_malmo_mcp_server/requirements.txt
# Core server dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
pydantic==2.5.3

# LLM API clients
anthropic==0.18.1
google-generativeai==0.3.2

# Malmo (will be installed from cloned repository)
# MalmoEnv will be added as git submodule

# Database for benchmarking
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
asyncio==3.4.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
